Added a new menu action "Insert Current Date/Time" (keyboard shortcut F5).
It's in the new "Insert Bubble or Text" submenu of the Edit menu.

If no bubble or link is selected, a new bubble containing the current date/time will be created.

Otherwise, the current/date/time will be added to the label of the selected bubble or link, either at the end of the label, or wherever the caret currently is.

The date/time format looks like this:

    2017-01-23 14:55:49